feat: enhance Lavalink integration with voice state management and command updates

+++ src/app/events/listeners/raw.ts
@@ -17,32 +17,27 @@ export class Raw extends BaseDiscordEvent {
     super(client, Events.Raw)
   }
 
-  run(packet: RawPacket) {
+  async run(packet: RawPacket) {
     const lavalink = this.client.lavalink
     const clientId = this.client.config.applicationId
 
     switch (packet.t) {
       case GatewayDispatchEvents.VoiceStateUpdate: {
         const { guild_id, channel_id, self_deaf, self_mute, user_id, session_id } = packet.d
-        if (!guild_id || !user_id) return
+        if (!guild_id || !user_id || clientId !== user_id) return
 
         if (channel_id === null) {
+          // TODO: maybe destroy the player here
           lavalink.deleteVoiceState(user_id)
-
-          if (clientId === user_id) {
-            // TODO: maybe destroy the player here
-            lavalink.deleteVoiceServer(guild_id)
-          }
-
           return
         }
 
-        lavalink.updateVoiceState(user_id, {
+        await lavalink.updateVoiceState(guild_id, {
           guildId: guild_id,
+          sessionId: session_id,
           voiceChannelId: channel_id,
           selfDeaf: self_deaf,
-          selfMute: self_mute,
-          sessionId: session_id
+          selfMute: self_mute
         })
         break
       }
@@ -51,7 +46,7 @@ export class Raw extends BaseDiscordEvent {
         const { token, guild_id, endpoint } = packet.d
         if (!guild_id || !token || !endpoint) return
 
-        lavalink.updateVoiceServer(guild_id, { token, endpoint })
+        await lavalink.updateVoiceServer(guild_id, { token, endpoint })
         break
       }
